refactor: split host file manager bookmark routes
